+++ src/constants/store.ts
@@ -3,9 +3,9 @@ import type {ExtractStoreData, SAFE_ANY} from '@helpers/type';
 import {getBetaVersion} from '@helpers/beta';
 import {type Components, getLatestVersion} from 'src/scripts/helpers';
 
-import {NEXTUI_CLI, NEXT_UI} from './required';
+import {HEROUI_CLI, HERO_UI} from './required';
 
-export type NextUIComponentsMap = Record<string, Components[0]>;
+export type HeroUIComponentsMap = Record<string, Components[0]>;
 
 export type Store = {
   debug: boolean;
@@ -14,21 +14,21 @@ export type Store = {
   latestVersion: string;
   betaVersion: string;
 
-  // NextUI
-  nextUIComponents: Components;
-  nextUIComponentsKeys: string[];
-  nextUIcomponentsPackages: string[];
-  nextUIComponentsKeysSet: Set<string>;
-  nextUIComponentsMap: NextUIComponentsMap;
-  nextUIComponentsPackageMap: NextUIComponentsMap;
+  // HeroUI
+  heroUIComponents: Components;
+  heroUIComponentsKeys: string[];
+  heroUIcomponentsPackages: string[];
+  heroUIComponentsKeysSet: Set<string>;
+  heroUIComponentsMap: HeroUIComponentsMap;
+  heroUIComponentsPackageMap: HeroUIComponentsMap;
 
-  // Beta NextUI
-  betaNextUIComponents: Components;
-  betaNextUIComponentsKeys: string[];
-  betaNextUIcomponentsPackages: string[];
-  betaNextUIComponentsKeysSet: Set<string>;
-  betaNextUIComponentsMap: NextUIComponentsMap;
-  betaNextUIComponentsPackageMap: NextUIComponentsMap;
+  // Beta HeroUI
+  betaHeroUIComponents: Components;
+  betaHeroUIIComponentsKeys: string[];
+  betaHeroUIcomponentsPackages: string[];
+  betaHeroUIComponentsKeysSet: Set<string>;
+  betaHeroUIComponentsMap: HeroUIComponentsMap;
+  betaHeroUIComponentsPackageMap: HeroUIComponentsMap;
 };
 
 /* eslint-disable sort-keys-fix/sort-keys-fix, sort-keys */
@@ -39,19 +39,19 @@ export const store = {
   latestVersion: '',
   betaVersion: '',
 
-  betaNextUIComponents: [],
-  betaNextUIComponentsKeys: [],
-  betaNextUIComponentsKeysSet: new Set(),
-  betaNextUIComponentsMap: {},
-  betaNextUIComponentsPackageMap: {},
-  betaNextUIcomponentsPackages: [],
+  betaHeroUIComponents: [],
+  betaHeroUIIComponentsKeys: [],
+  betaHeroUIComponentsKeysSet: new Set(),
+  betaHeroUIComponentsMap: {},
+  betaHeroUIComponentsPackageMap: {},
+  betaHeroUIcomponentsPackages: [],
 
-  nextUIComponents: [],
-  nextUIComponentsKeys: [],
-  nextUIComponentsKeysSet: new Set(),
-  nextUIComponentsMap: {},
-  nextUIComponentsPackageMap: {},
-  nextUIcomponentsPackages: []
+  heroUIComponents: [],
+  heroUIComponentsKeys: [],
+  heroUIComponentsKeysSet: new Set(),
+  heroUIComponentsMap: {},
+  heroUIComponentsPackageMap: {},
+  heroUIcomponentsPackages: []
 } as Store;
 /* eslint-enable sort-keys-fix/sort-keys-fix, sort-keys */
 
@@ -64,15 +64,15 @@ export async function getStore<T extends StoreKeys = StoreKeys>(
 
   if (!data) {
     if (key === 'latestVersion') {
-      data = (await getLatestVersion(NEXT_UI)) as SAFE_ANY;
+      data = (await getLatestVersion(HERO_UI)) as SAFE_ANY;
 
       store[key] = data;
     } else if (key === 'cliLatestVersion') {
-      data = (await getLatestVersion(NEXTUI_CLI)) as SAFE_ANY;
+      data = (await getLatestVersion(HEROUI_CLI)) as SAFE_ANY;
 
       store[key] = data;
     } else if (key === 'betaVersion') {
-      data = (await getBetaVersion(NEXT_UI)) as SAFE_ANY;
+      data = (await getBetaVersion(HERO_UI)) as SAFE_ANY;
 
       store[key] = data;
     }