+++ src/constants/templates.ts
@@ -1,8 +1,8 @@
 import type {CheckType} from '@helpers/check';
 
-export const APP_REPO = 'https://codeload.github.com/nextui-org/next-app-template/tar.gz/main';
-export const PAGES_REPO = 'https://codeload.github.com/nextui-org/next-pages-template/tar.gz/main';
-export const VITE_REPO = 'https://codeload.github.com/nextui-org/vite-template/tar.gz/main';
+export const APP_REPO = 'https://codeload.github.com/frontio-ai/next-app-template/tar.gz/main';
+export const PAGES_REPO = 'https://codeload.github.com/frontio-ai/next-pages-template/tar.gz/main';
+export const VITE_REPO = 'https://codeload.github.com/frontio-ai/vite-template/tar.gz/main';
 
 export const APP_DIR = 'next-app-template-main';
 export const PAGES_DIR = 'next-pages-template-main';
@@ -11,29 +11,29 @@ export const VITE_DIR = 'vite-template-main';
 export const APP_NAME = 'next-app-template';
 export const PAGES_NAME = 'next-pages-template';
 export const VITE_NAME = 'vite-template';
-export const DEFAULT_PROJECT_NAME = 'nextui-app';
+export const DEFAULT_PROJECT_NAME = 'heroui-app';
 
 export function tailwindTemplate(type: 'all', content?: string): string;
 export function tailwindTemplate(type: 'partial', content: string): string;
 export function tailwindTemplate(type: CheckType, content?: string) {
   if (type === 'all') {
     return `// tailwind.config.js
-const {nextui} = require("@nextui-org/react");
+const {heroui} = require("@heroui/react");
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
-    "./node_modules/@nextui-org/theme/dist/**/*.{js,ts,jsx,tsx}",
+    "./node_modules/@heroui/theme/dist/**/*.{js,ts,jsx,tsx}",
   ],
   theme: {
     extend: {},
   },
   darkMode: "class",
-  plugins: [nextui()],
+  plugins: [heroui()],
 };`;
   } else {
     return `// tailwind.config.js
-const {nextui} = require("@nextui-org/theme");
+const {heroui} = require("@heroui/theme");
 
 /** @type {import('tailwindcss').Config} */
 module.exports = {
@@ -44,7 +44,7 @@ module.exports = {
     extend: {},
   },
   darkMode: "class",
-  plugins: [nextui()],
+  plugins: [heroui()],
 };`;
   }
 }