+++ src/index.ts
@@ -24,12 +24,12 @@ import {compareVersions, getComponents} from './scripts/helpers';
 
 const commandList: CommandName[] = ['add', 'env', 'init', 'list', 'upgrade', 'doctor', 'remove'];
 
-const nextui = new Command();
+const heroui = new Command();
 
-nextui
-  .name('nextui')
+heroui
+  .name('heroui')
   .usage('[command]')
-  .description(getCommandDescAndLog(`\nNextUI CLI v${pkg.version}\n`, ''))
+  .description(getCommandDescAndLog(`\nHeroUI CLI v${pkg.version}\n`, ''))
   .version(pkg.version, '-v, --version', 'Output the current version')
   .helpOption('-h, --help', 'Display help for command')
   .allowUnknownOption()
@@ -60,11 +60,11 @@ nextui
     }
 
     if (!isArgs) {
-      const helpInfo = (await getCacheExecData('nextui --help')) as string;
+      const helpInfo = (await getCacheExecData('heroui --help')) as string;
 
       let helpInfoArr = helpInfo.split('\n');
 
-      helpInfoArr = helpInfoArr.filter((info) => info && !info.includes('NextUI CLI v'));
+      helpInfoArr = helpInfoArr.filter((info) => info && !info.includes('HeroUI CLI v'));
       // Add command name color
       helpInfoArr = helpInfoArr.map((info) => {
         const command = info.match(/(\w+)\s\[/)?.[1];
@@ -81,7 +81,7 @@ nextui
     process.exit(0);
   });
 
-nextui
+heroui
   .command('init')
   .description('Initializes a new project')
   .argument('[projectName]', 'Name of the project to initialize')
@@ -89,7 +89,7 @@ nextui
   .option('-p --package [string]', 'The package manager to use for the new project', 'npm')
   .action(initAction);
 
-nextui
+heroui
   .command('add')
   .description('Adds components to your project')
   .argument('[components...]', 'Names of components to add')
@@ -102,7 +102,7 @@ nextui
   .option('-b --beta [boolean]', 'Add beta components', false)
   .action(addAction);
 
-nextui
+heroui
   .command('upgrade')
   .description('Upgrades project components to the latest versions')
   .argument('[components...]', 'Names of components to upgrade')
@@ -112,7 +112,7 @@ nextui
   .option('-b --beta [boolean]', 'Upgrade beta components', false)
   .action(upgradeAction);
 
-nextui
+heroui
   .command('remove')
   .description('Removes components from the project')
   .argument('[components...]', 'Names of components to remove')
@@ -122,19 +122,19 @@ nextui
   .option('--prettier [boolean]', 'Apply Prettier formatting to the added content')
   .action(removeAction);
 
-nextui
+heroui
   .command('list')
   .description('Lists all components, showing status, descriptions, and versions')
   .option('-p --packagePath [string]', 'Specify the path to the package.json file')
   .option('-r --remote', 'List all components available remotely')
   .action(listAction);
-nextui
+heroui
   .command('env')
   .description('Displays debugging information for the local environment')
   .option('-p --packagePath [string]', 'Specify the path to the package.json file')
   .action(envAction);
 
-nextui
+heroui
   .command('doctor')
   .description('Checks for issues in the project')
   .option('-p --packagePath [string]', 'Specify the path to the package.json file')
@@ -145,7 +145,7 @@ nextui
   .option('-cp --checkPnpm [boolean]', 'Check for Pnpm', true)
   .action(doctorAction);
 
-nextui.hook('preAction', async (command) => {
+heroui.hook('preAction', async (command) => {
   const args = command.args?.[0];
   const options = (command as SAFE_ANY).rawArgs.slice(2);
   const noCache = options.includes('--no-cache');
@@ -160,11 +160,11 @@ nextui.hook('preAction', async (command) => {
 
   if (args && commandList.includes(args as CommandName)) {
     // Before run the command init the components.json
-    const nextUIComponents = (await getComponents()).components;
-    const nextUIComponentsBeta = (await getComponents()).betaComponents;
+    const heroUIComponents = (await getComponents()).components;
+    const heroUIComponentsBeta = (await getComponents()).betaComponents;
 
-    initStoreComponentsData({beta: false, nextUIComponents});
-    store.beta && initStoreComponentsData({beta: true, nextUIComponents: nextUIComponentsBeta});
+    initStoreComponentsData({beta: false, heroUIComponents: heroUIComponents});
+    store.beta && initStoreComponentsData({beta: true, heroUIComponents: heroUIComponentsBeta});
   }
 
   const [cliLatestVersion, latestVersion] = await Promise.all([
@@ -176,7 +176,7 @@ nextui.hook('preAction', async (command) => {
   store.latestVersion = latestVersion;
   store.cliLatestVersion = cliLatestVersion;
 
-  // Add NextUI CLI version check preAction
+  // Add HeroUI CLI version check preAction
   const currentVersion = pkg.version;
 
   if (compareVersions(currentVersion, cliLatestVersion) === -1) {
@@ -188,18 +188,18 @@ nextui.hook('preAction', async (command) => {
         `Available upgrade: v${currentVersion} -> ${chalk.greenBright(
           `v${cliLatestVersion}`
         )}\nRun \`${chalk.cyan(
-          'npm install -g nextui-cli@latest'
+          'npm install -g heroui-cli@latest'
         )}\` to upgrade\nChangelog: ${chalk.underline(
-          'https://github.com/nextui-org/nextui-cli/releases'
+          'https://github.com/frontio-ai/heroui-cli/releases'
         )}`
       )}`,
-      title: gradientString('NextUI CLI')
+      title: gradientString('HeroUI CLI')
     });
     Logger.newLine();
   }
 });
 
-nextui.parseAsync(process.argv).catch(async (reason) => {
+heroui.parseAsync(process.argv).catch(async (reason) => {
   Logger.newLine();
   Logger.error('Unexpected error. Please report it as a bug:');
   Logger.log(reason);